requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
flake8==7.0.0
mccabe==0.7.0
pillow==10.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
sqlparse==0.4.4
tzdata==2024.1
